Add databinding to register and login XML files, also refactor RegisterViewModel and added a Factory class to instantiate

+++ app/src/main/java/com/example/food_notes/db/dbviews/UserDetails.java
@@ -4,7 +4,7 @@ import androidx.room.DatabaseView;
 
 @DatabaseView("SELECT * FROM users")
 public class UserDetails {
-    public Long user_id;
+    public Integer user_id;
     public String username;
     public String password;
 }
